mints membership

+++ src/interfaces/iInstanceDAO.sol
@@ -6,30 +6,27 @@ interface iInstanceDAO {
 
     function giveOwnership() external returns (address);
 
-    function wrapMint(uint256 amount) external returns(bool);
+    function wrapMint(uint256 amount) external returns (bool);
 
     function unwrapBurn(uint256 amount_) external returns (bool);
 
-    function incrementSubDAO() external returns (uint);
+    function incrementSubDAO() external returns (uint256);
 
     /// view
 
     function owner() external view returns (address);
-    
+
     function internalTokenAddr() external view returns (address);
 
     function baseTokenAddress() external view returns (address);
 
-    function baseID() external view returns (uint);
+    function baseID() external view returns (uint256);
 
-    function localID() external view returns (uint);
+    function localID() external view returns (uint256);
 
     function mintMembershipToken(address to_) external returns (bool);
     /// only owner
     function setPerSecondInterestRate(uint256 ratePerSec) external returns (bool);
 
     function giveOwnership(address newOwner_) external returns (address);
-
-
-
 }
